Basic simulation of physical lab
Creates 3 containers to simulate the situation. 2 receivers and 1 transmitter. Each contains a small python script for their parts.
The transmitter is multicasting to the network and each receiver is subscribing to these messages and currently just printing them.

The physical nodes and the network are managed with Docker Compose and there are some simple orchestration commands added to the Makefile.

Additionally each python script is wrapped in an init script to allow it to be managed as a service on the container which
we'll use later once Ansible is hopping in and controlling the workflow.