Working commit of Ansible connecting from Orchestrator
Lab is now mostly stubbed out with an Orchestration container running the ansible.
The orchestrator can take its default key and connect out to all of the hosts.
The reason we need this orchestrator is due to some intended behaviors with Docker for Mac. TL;DR it is extremely difficult to
get the Mac host machine to be able to route directly to the virtual docker network, so rather than deal with that...just
run the tests from the Orchestrator.

Re-organized the repo a little bit to move things under the lab-setup dir.
Most provided files are now under the setup for the orchestrator container.