Moved lab playbooks to orchestrator
All ansible related to starting the various services has been moved to the Orchestrator container.
There is now a new playbook in the root of the project responsible for telling the Orchestrator to start/stop the services

Makefile has been updated to start the lab and tell the orchestrator to transmit data across the topology.